+++ scripts/src/commands/e2e.ts
@@ -1,6 +1,90 @@
+import * as path from 'node:path'
 import { cac } from 'cac'
-import { getRepoRoot, listAppDirectories } from '../helpers/repo.ts'
-import { runE2E, type E2EOptions } from '../helpers/e2e.ts'
+import { $ } from 'zx'
+import {
+	assertAppExists,
+	getRepoRoot,
+	listAppDirectories,
+} from '../helpers/repo.ts'
+
+const E2E_COMPOSE_FILE = path.resolve(
+	import.meta.dirname,
+	'..',
+	'e2e.compose.yaml',
+)
+
+export type E2EOptions = {
+	local?: boolean
+	preview?: string
+	staging?: boolean
+	production?: boolean
+	all?: boolean
+}
+
+function getMode(
+	options: E2EOptions,
+): 'local' | 'preview' | 'staging' | 'production' | null {
+	const modes = [
+		options.local,
+		options.preview,
+		options.staging,
+		options.production,
+	].filter(Boolean)
+	if (modes.length > 1) {
+		console.error(
+			'Error: Specify only one of --local, --preview, --staging, --production',
+		)
+		process.exit(1)
+	}
+
+	if (options.local) return 'local'
+	if (options.preview) return 'preview'
+	if (options.staging) return 'staging'
+	if (options.production) return 'production'
+
+	return 'local'
+}
+
+function getTargetUrl(app: string, options: E2EOptions): string {
+	if (options.preview) {
+		return `https://${app}-pr-${options.preview}.ariaamini.com`
+	}
+	if (options.staging) {
+		return `https://${app}.staging.ariaamini.com`
+	}
+	if (options.production) {
+		return `https://${app}.ariaamini.com`
+	}
+	return ''
+}
+
+async function runE2E(
+	repoRoot: string,
+	app: string,
+	options: E2EOptions,
+): Promise<void> {
+	const mode = getMode(options)
+	const targetUrl = getTargetUrl(app, options)
+
+	assertAppExists(repoRoot, app)
+
+	const interactive = $({
+		cwd: repoRoot,
+		stdio: 'inherit',
+		env: {
+			...process.env,
+			APP_NAME: app,
+			BASE_URL: targetUrl,
+			CI: targetUrl || process.env.CI === 'true' ? 'true' : '',
+		},
+	})
+
+	if (mode === 'local') {
+		await interactive`docker compose -f ${E2E_COMPOSE_FILE} run --build --rm e2e`
+	} else {
+		throw new Error(`e2e --${mode} not yet implemented`)
+	}
+}
 
 export function createE2ECommand(): ReturnType<typeof cac> {
 	const cli = cac('aamini e2e')
